lfs{api,http}: test bare repo local path endpoints
When generating a local path endpoint, we now test whether the
path corresponds to a bare repo lacking a .git directory, and if so,
we ensure the endpoint path has any trailing .git segment trimmed.

Conversely, we ensure the endpoint path has a final .git segment if
a .git directory is found.

The test suite is updated to check the four relevant cases.