Makefile: build macOS release assets as zip file
In the future, we're going to want to sign and notarize our macOS
binaries.  When we do that, we're going to need to produce a zip file,
since we can't notarize a tarball but we can notarize a zip file.  In
order to make the release assets as similar to the ones that we build in
development, let's ship the Darwin assets as a zip file instead of a
tarball.