CI: prune unused Docker images after use
At the moment, we're running out of disk space when building packages in
Docker containers.  This is because our CI VMs have only 14 GB of
storage.  To help deal with this problem, let's add an option to the
run_dockers.bsh script to prune images after using them so we can free
some of that disk space and let the remaining jobs complete.