t/testenv: set LC_ALL to C
When we run the test suite, we search for specific output in Git's
output, but this output is dependent on the locale.  Consequently,
running the test suite in a language for which Git has translations,
such as Spanish, French, or Russian, can cause test failures.

Let's set the locale to C, which will result in English output, so that
the tests pass.  Ideally we'd use C.UTF-8, but not all systems have that
locale, so we need to stick with C.