git: consider full refspec when determining seen refs
We avoid rewriting references that we've already seen, since in some
cases we will need to order references to properly handle tags pointing
to tags.  However, we use ref.Name for this, which is the short name,
not the full refspec.  Consequently, if we've already seen a ref with
the same short name (e.g., a branch when we're processing a tag), we end
up skipping the second ref.

This is not what we want, so let's use the full refspec when populating
the set of items we've already seen.