git: update regex for SHA-256
With SHA-256, the length of a Git object ID in hex is 64 characters, not
40.  Let's match either possible value with a non-capturing group and
use a variable to store this value so if we ever need to update it
again, it only requires one change.  Use the constants we've defined so
it's obvious what the values mean.

While we're here, we limit the output to only lowercase hex characters,
since Git will only ever produce those values (and not uppercase
values), and we're parsing its output here.