t: add tests for --worktree install and uninstall
We add a number of tests, conditional on the Git version, to
validate the functionality of the --worktree option when
installing or uninstalling.

Note that because this simply passed through to the "git config"
command, we inherit the behaviour that --worktree functions
like --local when only a single, default working tree is in use.

And if we have a Git version older than 2.20.0, we confirm that
the --worktree option is not supported.