t: mark files we're locking as lockable
Currently our unlock tests don't mark the files we're operating on as
lockable.  In most of our tests this won't matter, but there are some
that will change in a future commit, so let's set up most of our test
repositories such that all files ending in ".dat" are lockable.