git: add lookup function for empty tree
There are several places we need to look up an empty tree.  When we were
supporting only SHA-1, we could hard-code this value, but we can't do
that now.  Since our codebase doesn't know about the hash algorithm in
use (and really doesn't need to), let's just ask Git to provide us with
the hash of the empty tree and cache it.

We use a mutex for thread safety, even though right now we don't need
it, because someone may add an additional use of this in the future.