git: improve performance of remote ref listing
When repositories have a large number (hundreds of thousands) of refs, Git
LFS adds noticeable overhead on the time taken by `git show-ref` and
`git ls-remote`. To reduce this overhead:

* Avoid regular expressions when parsing lines of Git command output.
  Regexp evaluation was the most expensive part of this process in
  profiling.

* Minimize the number of loops when computing skipped refs by combining
  the missing and skipped checks and converting the list of remote refs
  into a pre-filtered set.

In a repository with ~300k remote references, these changes reduce the
time spent in the pre-push command from 8.26s to 6.31s.