{README,SECURITY}.md: add security bug report docs
Following the GitHub template for adding security policies for
open-source projects, we add a SECURITY.md file which describes
the relevant policy for the Git LFS client, as well as providing
links to related resources and general information.

In order to facilitate reports of security vulnerabilities via
secure email, we also add to the core team roster on our home
page links to the PGP keys of each active team member.

See also https://docs.github.com/en/github/managing-security-vulnerabilities/adding-a-security-policy-to-your-repository