ci: don't reinstall openssl with Homebrew
GitHub Actions ships its own version of the openssl Homebrew package
that fails to reinstall with an EPERM error.  Since the reason we did
these contortions was to relink gettext and we finally gave up and
disabled gettext, let's just go back to a brew link --force and stop
trying to reinstall.