subprocess: don't initialize environment at startup
Currently, the subprocess package reads from the environment as it's
created at startup when the init function is called.  However, we'll
soon want to modify the environment in this case before it gets
processed, so let's change the code to use a mutex to initialize the
environment once before using it and simply call that before using the
environment we've set up.

We'll want to reset the environment as well in a future commit, so let's
be sure to add a function for that.  We reuse the same internal function
and just ignore the return value to make our code paths simpler.