config: map missing port to default for HTTP key lookups
When Git performs a lookup for a key with urlmatch semantics, it matches
the default port if none is specified in the URL.  However, our code
currently doesn't do that, instead having a missing port match only
another missing port entry.

Let's fix the behavior to match Git's by mapping a missing port to the
default port.  We add tests for checking both URLs (the config one and
the one we're sending data to) to make sure that our code is robust.