git: match gitattributes patterns strictly
When we look for patterns in the .gitattributes file, we want to match
these patterns strictly: that is, exactly the way that Git does.  We
don't want to use a different algorithm because that will lead to
confusing behavior when our code behaves differently from Git does.

Let's fix this by using our new strict matching when reading patterns
from the gitattributes files and add a test to make sure we don't
regress anything.