Drop support for Debian 8
Debian 8 went out of support on June 30, 2020.  Ubuntu 14.04, which uses
the same packages, went out of support in 2019.

Since there's no longer any free security support and we cannot
reasonably be expected to run insecure code in CI, since that's
dangerous to others on the Internet, let's drop support for Debian 8.