t/t-push.sh: revise test of truncated object ID
We adjust the test of invalid object ID handling in "git lfs push"
added in commit 3f4662b489a1bb84fda605472d4e662fc6abb0c9 so
that it passes again now that its logic is fully executed, as it
was previously short-circuited by the bug in push_all_setup()
(described in commit f9381fa37cb25af17c6d90f4eea054770cb5692d)
and with that bug fixed, the test actually fails.

The test intends to check for an error message from "git lfs push"
when passed an OID shorter than 4 characters, so we revise it to
generate a 3-character OID; that then allows the test to pass.

We also adjust the name passed to push_all_setup() to avoid
conflicting with the "everything" test of the --all option.