t/t-migrate-import: check for correct missing object
It was clearly intended that we be checking for a pointer not to exist
with the contents of a.txt, not a.md, at this point.  This passes either
way because the file isn't a pointer at all and therefore doesn't match
anything, but it's better to check for the proper thing regardless.