.github/workflows: don't try to install older Git
Git for Windows 2.27.0 had a bug where symlinks would appear perpetually
modified when core.fscache was enabled, causing our CI job to fail
because it looks for modified files after running the formatter.
However, a newer version of Git is now available and our old code is
failing, so let's remove it and use the newer version of Git that's
shipped by default.