commands: change default branch name
Since the default branch name has already changed for GitHub and is
likely to change for Git in the future, let's suggest people use the new
default branch name, main.  The example is arbitrary and doesn't matter
much, but picking what is likely to be the default will be less
confusing.