lfsapi: don't warn about duplicate but identical aliases
Currently, we warn about aliases (e.g., url.*.insteadOf) when there's
already an alias for the same URL; that is, when the insteadOf entry
matches an existing one.  However, in some environments, multiple
aliases are defined, but all to the same root URL (the one in the key).

In such a case, don't warn, since these aliases are already identical.
Do continue to warn if the root URL is different at all, since these are
different aliases for the same URL, and add a test for both of these
cases.