Gently steer users to discussions
Discussions provide some nice features for questions, including a way to
mark suggested answers and threading.  Let's encourage users to use them
for things which are not bug reports or feature requests.

Let's also mention the FAQ in the CONTRIBUTING documentation, so that
users who are about to start a discussion but have a common question can
find their answer quickly.  The README already has such an entry
directly above the text we've added.