lfs: make all blob size checks consistent
The blobSizeCutoff limit is defined such that "any file
with a size below this cutoff will be scanned" when looking
for LFS pointer blobs.  In general this is true, but there
are few instances where the check is applied such that blobs
matching the cutoff limit are also scanned.

We therefore adjust these to match the same comparison logic
used elsewhere with this limit.

Two of these checks are in the PointerScanner.next() method
used by the runCatFileBatch() function; they were introduced
in commit 844c0b0db2bf3d39f52773e5f63ee10683a28342 in
PR #2070 and scan Git objects for eligible LFS pointers
when reading the output of "git cat-file --batch".  Note
that the corresponding runCatFileBatchCheck() function's
catFileBatchCheckScanner.next() method excludes blobs whose
size is above or equal to the limit (with a default limit of
blobSizeCutoff); this aligns with other usage and checks
of the blobSizeCutoff maximum as well.

The other check is in the DecodePointerFromFile()
function where it is used to exclude from consideration as
Git LFS pointers any files larger than blobSizeCutoff.
This function is ultimately called by the "git lfs checkout"
command; the original implementation was added in commit
001ddcd7cfae528a6594be9bab4562c6659a00e7 in PR #527.