Makefile: build ARM64 targets for Darwin
With macOS Big Sur, ARM64 is now a supported architecture for Darwin.
Building binaries for this case is supported since Go 1.16.  Since users
would like to have native signed binaries for macOS, let's go ahead and
build Darwin packages as part of the make release command.

Note that this will necessarily require users to have Go 1.16 available
to run that command, since otherwise the attempt to build for ARM64
Darwin will fail.  Since this is mostly run in CI and on developer
machines, that's okay.