t: reorder prune stash checks to match actions
We slightly reorder the assertions following the
"git stash" command in several of our tests of the
"git lfs prune" command so that the assertions are
made in the same order as the stashed modifications.
This will ease the addition of more assertions in
subsequent commits.

We also drop the unnecessary cleanup of the stash
and working tree at the end of one test.