commands,t: make import default to saving work
We update the prompt to be more explicit when the import
(or export) migration commands notice a "dirty" working
tree with uncommitted changes, and we also revise the
default action in this case to be to exit early and
preserve the work in progress.

h/t bk2204 for the suggested revisions.