Add config variables for default remotes
Currently, if there's no default remote for the current branch set, we
look for the name of any single remote, and if not, we fall back to
"origin."

However, some people prefer to use a different name for their remotes,
such as "upstream".  We don't really want to mandate that users provide
a specific name for their remotes, so let's add two variables,
remote.lfsdefault and remote.lfspushdefault, that control the default
remote and the default push remote, respectively.  Let's document them,
and the lookup order, as well, so folks can reason about the behavior
adequately.