errors: allow reporting invalid pointers
Currently, Git LFS tries very hard to ignore Git files that are not
pointers.  However, we'd like to add a mode to git lfs fsck where people
can scan for invalid pointers so they can check for problems in CI.
Let's add an error type that we can use to report these errors and
extract the relevant data from them to provide to the user.