t/t-migrate-info.sh: add --pointers option tests
We add a number of tests of the functionality of the new
--pointers option to the "migrate info" command in order
to confirm it works as expected in all modes, and both with
and without Git LFS pointers in the test repositories.

We also add tests of repositories where some file extensions
are tracked in .gitattributes files and the corresponding blobs
should be Git LFS pointers, but are not yet.  In these cases,
we do want to report these file extensions in summary data of
the output listing.

In order to ensure our tests succeed on Windows, we alter
several fixture setup functions to create the .gitattributes
files using "echo" instead of "git lfs track".  This keeps
the line endings consistent across platforms, and therefore
the .gitattributes file sizes in the "migrate info" summaries.