lfsapi: move SSHTransfer instantiation to lfsapi client
The lfsapi client is used to perform operations for SSH authentication
already, so we know we'll have one wherever SSH operations might be
done.  Let's move the instantiation to the client so that we can reuse
it both in the transfer queue code and in the locking code as well.