locking: implement a generic transport adapter
The locking client takes a remote argument, which means that each
individual operation can operate on a different remote.  However, the
SSH transport adapter wants to preserve a single long-lived connection
to a single remote, which is better for users who use interactive SSH
login methods.

Let's implement a single generic transport adapter which caches the
operations and remotes that it's used for and reuses the same transport
adapter if so.  This avoids needlessly spawning multiple processes.

Note that this wires up the pure SSH transport adapter at the same time.