t: test locking with git-lfs-transfer
Test locking using git-lfs-transfer as well as git-lfs-authenticate, and
label the tests to distinguish between them.  Add functions to assert
and refute the existence of a lock over SSH and, to implement these, to
speak basic pkt-line commands over a pipe.

While we're at it, remove a debugging statement in one of the tests.