commands,docs,t: add migrate info --fixup option
We add a --fixup option to the "git lfs migrate info"
command which has similar properties as the --fixup option
to the "git lfs migrate import" command, namely that it
looks for files in the repository which should be Git LFS
pointers (according to the filter settings in .gitattributes)
but which are not.

In the case of the "migrate info" command this new option
only reports the sum totals for any such files, and does
not make any changes to the repository.  When the --fixup
option is supplied, the "migrate info" command will not
output a summary of any other files, including .gitattributes
files, which are otherwise normally included in the command's
final output.

The new --fixup option is taken to imply the --pointers=ignore
option, and is incompatible with any other settings of that
option.