docs: add support for hash algorithm detection
While SHA-256 is presently considered strong, it might not always be, so
we should consider supporting other hash algorithms.  We anticipate
that, much like Git, exactly one hash algorithm will be in use at a
time per repository.  To make it easier for the client and server to
negotiate a suitable algorithm, let's add a field to designate the hash
algorithm in batch requests.

This allows the client to declare to the server the hash algorithm on
first upload, and the server can respond with the hash algorithm in use
(if it supports multiple) or a 409 response if it cannot handle that.
We expect clients and servers to both gracefully handle the absence of
this value and assume SHA-256 if not specified.