tq: pass hash algorithm during batch requests
During a batch request, pass the hash algorithm we're using to the
remote server, and read the value, if any, that we get back.  If it is
not either absent or the string "sha256", fail, since that means that
the client and server don't agree on the proper hash algorithm.